pyproject.toml
[project]
name = "isl-voice-backend"
version = "0.1.0"
description = "ISL Sign Language to Voice — Vision Agents Backend"
requires-python = ">=3.12"
dependencies = [
    "vision-agents[getstream,gemini,elevenlabs,deepgram]",
    "roboflow>=0.2.0",
    "python-dotenv>=1.0.0",
    "stream-chat>=4.0.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
